requirements.txt
absl-py==2.1.0
astunparse==1.6.3
bctpy==0.6.1
blessed==1.20.0
DCLS==0.1.0
docker-pycreds==0.4.0
flatbuffers==24.3.25
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
google-pasta==0.2.0
gpustat==1.1
grpcio==1.62.2
importlib_metadata==8.7.0
keras==3.3.2
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==11.525.112
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
opt-einsum==3.3.0
optree==0.11.0
packaging==21.3
pathlib==1.0.1
pillow==11.3.0
Pygments==2.17.2
rich==13.7.1
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
scipy==1.13.1
matplotlib==3.5.1
sentry-sdk==1.35.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
tqdm==4.67.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
torch==1.13.0
triton==2.1.0
typing_extensions==4.11.0
urllib3==1.26.18
wandb==0.16.0
Werkzeug==3.1.3
zipp==3.23.0
